+++ docs/changelog.md
@@ -23,6 +23,7 @@
         * ENABLE_PATTERN_CREATION_FOR_RECIPES_WITH_REQUIREMENTS
     * Vanilla recipes can now be transferred from JEI with the `+` button if the table's vanilla recipe feature is enabled in the config [(#216)](https://github.com/codetaylor/artisan-worktables/issues/216)
     * CraftTweaker marks are now available for secondary ingredients and tools [(#219)](https://github.com/codetaylor/artisan-worktables/issues/219)
+    * Crafting Tweaks integration [(#80)](https://github.com/codetaylor/artisan-worktables/issues/80)
 
 ---